+++ backend/app/routers/apps.py
@@ -1,3 +1,24 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: MIT
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+# DEALINGS IN THE SOFTWARE.
+
 import asyncio
 import logging
 from typing import Annotated
@@ -10,7 +31,8 @@ from app.auth import admin_only, authenticated_only
 from app.models import (
     PublishedAppResponse, PublishedAppModel, PublishedApp, NvcfFunctionStatus
 )
-from app.nvcf import get_nvcf_functions, get_nvcf_function_status
+from app.nvcf import get_nvcf_functions, get_nvcf_function_status, \
+    nvcf_function_cache
 
 router = APIRouter()
 logger = logging.getLogger('uvicorn.error')
@@ -116,6 +138,8 @@ async def publish_app(app_id: str, app: PublishedApp):
     )
 
     result = await PublishedAppResponse.from_tortoise_orm(app_model)
+
+    nvcf_function_cache.clear()
     functions = await get_nvcf_functions()
 
     result.status = get_nvcf_function_status(
@@ -155,3 +179,5 @@ async def delete_app(app_id: str):
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"App {app_id} is not found.",
         )
+
+    nvcf_function_cache.clear()