+++ web/src/pages/Home.tsx
@@ -1,3 +1,26 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: MIT
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ */
+
 import { Card, Group, Loader, SimpleGrid, Stack, Title } from "@mantine/core";
 import { IconAppWindow, IconTableFilled } from "@tabler/icons-react";
 import { useQuery } from "@tanstack/react-query";
@@ -23,7 +46,7 @@ export default function Home() {
     isLoading: isLoadingApps,
     data: appsByPages,
     error: appError,
-  } = useQuery<Map<StreamingApp["page"], StreamingApp[]>>({
+  } = useQuery<Map<StreamingApp["page"], Set<StreamingApp>>>({
     queryKey: ["get-apps"],
     queryFn: async () => getStreamingApps({ config }),
   });
@@ -46,31 +69,30 @@ export default function Home() {
   const selectedPage = searchParams.get("page") ?? pageNames?.[0];
 
   const apps = selectedPage ? appsByPages?.get(selectedPage) : [];
-  const categories =
-    apps?.reduce(
-      (categories, app) => {
-        const categoryName = app.category ?? "";
-        const category = categories[categoryName] ?? [];
-        category.push(app);
-        categories[categoryName] = category;
-        return categories;
-      },
-      {} as Record<string, StreamingApp[]>
-    ) ?? {};
+  const categories = Array.from(apps?.values() ?? []).reduce(
+    (categories, app) => {
+      const categoryName = app.category ?? "";
+      const category = categories[categoryName] ?? [];
+      category.push(app);
+      categories[categoryName] = category;
+      return categories;
+    },
+    {} as Record<string, StreamingApp[]>,
+  );
 
   const error = appError || pageError;
   return (
     <Stack>
       <Header />
       <Stack px={"xl"} py={"md"}>
-        <Title c={"gray"}>Welcome to Omniverse on DGX Cloud!</Title>
+        <Title c={"gray"}>{config.userInterface.title}</Title>
         {isLoadingApps || isLoadingPages ? (
           <Loader />
         ) : error ? (
           <LoaderError title={"Failed to load streaming applications"}>
             {error.toString()}
           </LoaderError>
-        ) : appsByPages!.size ? (
+        ) : appsByPages?.size ? (
           <Group align={"start"} justify={"stretch"} wrap={"nowrap"}>
             <ApplicationPages pages={pageNames} selectedPage={selectedPage} />
             <Stack flex={1}>