+++ web/src/router.tsx
@@ -1,3 +1,26 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: MIT
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ */
+
 import { createBrowserRouter } from "react-router-dom";
 import AppInfo from "./pages/AppInfo";
 import Login from "./pages/Login";
@@ -9,6 +32,7 @@ import AppStream from "./pages/AppStream";
 import NucleusAuthenticate from "./pages/NucleusAuthenticate";
 import NucleusSSO from "./pages/NucleusSSO";
 import AppStreamList from "./pages/AppStreamList";
+import { DeepLink } from "./pages/DeepLink.tsx";
 
 export const router = createBrowserRouter([
   {
@@ -42,6 +66,10 @@ export const router = createBrowserRouter([
       {
         path: "app/:appId",
         element: <AppInfo />,
+      },
+      {
+        path: "deeplink",
+        element: <DeepLink />,
       },
       {
         path: "sessions",